Add spell checking
This adds a `make spelling` target that runs the spell checker.  Note that
the golang imports are unlikely to be found in the default word list, so we
have code to generate a word list based on our imports.

This does not include the spelling fixes (including contortions to feed the
spell checker); that will be in the next commit.

Signed-off-by: Mark Yen <mark.yen@suse.com>