Controllers: webhook: reuse CA key
When running with _some_ external controllers, we must use the same CA to
sign the webhook certificate.  For that, we need to save the CA key to disk
so it can be picked up and re-used for the external controller.  Given that
we already save the webhook private key to disk in the same location, this
is unlikely to be an issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>